+++ packages/core/src/cli.ts
@@ -328,7 +328,7 @@ function help(): void {
       --delay [seconds]:       set a delay in seconds before starting Matterbridge in the first 5 minutes from a reboot (default 120)
       --fixed_delay [seconds]: set a fixed delay in seconds before starting Matterbridge (default 120)
       --no-ansi:               disable ANSI color output in the logs    
-      --no-reset-sessions:     disable resetting of sessions and resumption records on shutdown 
+      --reset-sessions:        reset sessions and resumption records on shutdown (use only if your controller has issue reconnecting on restart)
   `);
   process.exit(0);
 }