+++ packages/core/src/matterbridge.device.server.test.ts
@@ -174,6 +174,7 @@ describe('Matterbridge Device serverMode=server', () => {
     const stopServerNodeSpy = jest.spyOn(matterbridge as any, 'stopServerNode');
 
     // Destroy the Matterbridge instance
+    process.argv.push('--reset-sessions');
     await destroyInstance(matterbridge, 0, 0);
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, `Cleanup completed. Shutting down...`);
     expect(stopServerNodeSpy).toHaveBeenCalledTimes(1);