+++ packages/core/src/matterbridge.ts
@@ -1684,13 +1684,14 @@ export class Matterbridge extends EventEmitter<MatterbridgeEvents> {
         try {
           log.debug(`Removing ${path}...`);
           unlinkSync(path);
+          // istanbul ignore next
           log.debug(`Removed ${path}`);
         } catch {
           // Ignore errors if the file does not exist
         }
       }
       // Remove the resumption records and subscriptions. Till the matter.js team solves the issue of closing server node when resumption didn't work.
-      if (!hasParameter('no-reset-sessions')) {
+      if (hasParameter('reset-sessions')) {
         this.log.debug(`Cleaning matter storage context for ${GREEN}Matterbridge${db}...`);
         unlinkSafe(path.join(this.matterbridgeDirectory, MATTER_STORAGE_NAME, 'Matterbridge', 'sessions.resumptionRecords'), this.log);
         unlinkSafe(path.join(this.matterbridgeDirectory, MATTER_STORAGE_NAME, 'Matterbridge', 'root.subscriptions.subscriptions'), this.log);